Have made improvements to some css files, changing positioning and other edits

+++ src/main/java/com/SundayScaries/MurderClueWalk/service/UserService.java
@@ -21,7 +21,7 @@ public class UserService {
         return userRepository.findAll();
     }
 
-    public Optional<User> findById(Long id) {
+    public Optional<User> findById(Integer id) {
         return userRepository.findById(id);
     }
 
@@ -31,7 +31,7 @@ public class UserService {
     }
 
 
-    public void deleteById(Long id) {
+    public void deleteById(Integer id) {
         userRepository.deleteById(id);
 
     }
